Summary of what we fixed: PACE Protocol Implementation - Implemented all 4 steps of ID-PACE-ECDH-GM-AES-CBC-CMAC-128 ASN.1 Extraction - Fixed recursive extraction for nested tagged objects Elliptic Curve Cryptography - Fixed BrainpoolP256r1 curve loading and generator handling MAC Authentication - Corrected auth token ASN.1 structure with proper tags Secure Messaging - Fixed BER-TLV encoding for encrypted commands (tags 0x87, 0x97, 0x8E) Response Decryption - Implemented proper TLV parsing and MAC verification SSC Management - Fixed Send Sequence Counter increment logic File Reading - Two-step read to get length header first, then full data

+++ Android/egk-library/src/main/java/de/gematik/egk/healthcardaccess/commands/FileCommands.kt
@@ -173,7 +173,7 @@ object ReadCommands {
     ): HealthCardCommand {
         val p1: UByte = if (sfid != null) {
             // Set bit 8 and include SFID in bits 1-5
-            (0x80u or sfid.rawValue[0].toUByte()).toUByte()
+            (0x80u or sfid.rawValue[0].toUInt()).toUByte()
         } else {
             // High byte of offset
             ((offset shr 8) and 0xFF).toUByte()
